Fixed (GCC 4.9.1, debug mode) compiler error.

+++ src/core/RDP/PersistentKeyListPDU.hpp
@@ -385,7 +385,7 @@ private:
             this->numEntriesCache[2] +
             this->numEntriesCache[3] +
             this->numEntriesCache[4],
-            MAXIMUM_ENCAPSULATED_BITMAP_KEYS
+            uint32_t(MAXIMUM_ENCAPSULATED_BITMAP_KEYS)
         );
     }
 };  // PersistentKeyListPDUData
